+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - **Dropdown** forward ref to select
+- **Pagination** Next button disabled
 
 ## [6.0.0] - 2018-08-23