+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [6.0.1] - 2018-08-23
+
 ### Fixed
 - **Dropdown** forward ref to select
 - **Pagination** Next button disabled